EdgeNode/pyproject.toml
[project]
name = "edge-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "paho-mqtt>=2.1.0",
    "requests>=2.32.3",
]
